+++ zface/src/utils/mod.rs
@@ -1,3 +1,5 @@
 pub mod print_keys;
 pub mod mnemonics;
+pub mod getter;
+
 pub use self::print_keys::*;
