+++ zface/src/wallet/commands.rs
@@ -3,9 +3,9 @@ use std::collections::HashMap;
 use crate::term::Term;
 use crate::derive::ChildIndex;
 use crate::utils::mnemonics::*;
+use crate::error::{Result, KeystoreError};
 use super::{WalletDirectory, KeystoreDirectory, DirOperations};
 use super::keyfile::{KeyFile, IndexFile};
-use super::error::{Result, KeystoreError};
 use super::config::*;
 use bip39::{Mnemonic, Language, MnemonicType, Seed};
 use rand::Rng;
